+++ src/peakrdl_rust/templates/components/addrmap.rs
@@ -4,15 +4,41 @@
 {{macros.includes(ctx)}}
 
 {{ctx.comment}}
-#[derive(Copy, Clone, Eq, PartialEq)]
-pub struct {{ctx.type_name|kw_filter}} {
+#[derive(Eq, PartialEq)]
+{% set struct_name = ctx.type_name|kw_filter %}
+pub struct {{struct_name}}<'io, IO = peakrdl_rust::io::PtrIO> {
     ptr: *mut u8,
+    io: &'io IO,
 }
 
-unsafe impl Send for {{ctx.type_name|kw_filter}} {}
-unsafe impl Sync for {{ctx.type_name|kw_filter}} {}
+unsafe impl<IO: Sync> Send for {{struct_name}}<'_, IO> {}
+unsafe impl<IO: Sync> Sync for {{struct_name}}<'_, IO> {}
 
-impl {{ctx.type_name|kw_filter}} {
+// manually implement Copy to ease generic bounds
+// (IO does not need to be Copy)
+impl<IO> Copy for {{struct_name}}<'_, IO> {}
+
+// manually implement Clone to ease generic bounds
+// (IO does not need to be Clone)
+impl<IO> Clone for {{struct_name}}<'_, IO> {
+    fn clone(&self) -> Self {
+        *self
+    }
+}
+
+impl {{struct_name}}<'static> {
+    /// # Safety
+    ///
+    /// The caller must guarantee that the provided address points to a
+    /// hardware register block implementing this interface.
+    #[inline(always)]
+    #[must_use]
+    pub const unsafe fn from_ptr(ptr: *mut ()) -> Self {
+        Self { ptr: ptr.cast::<u8>(), io: &peakrdl_rust::io::PtrIO }
+    }
+}
+
+impl<'io, IO> {{struct_name}}<'io, IO> {
     /// Size in bytes of the underlying memory
     pub const SIZE: usize = {{"0x{:_X}".format(ctx.size)}};
 
@@ -22,8 +48,8 @@ impl {{ctx.type_name|kw_filter}} {
     /// hardware register block implementing this interface.
     #[inline(always)]
     #[must_use]
-    pub const unsafe fn from_ptr(ptr: *mut ()) -> Self {
-        Self { ptr: ptr.cast::<u8>() }
+    pub const unsafe fn from_ptr_with(ptr: *mut (), io: &'io IO) -> Self {
+        Self { ptr: ptr.cast::<u8>(), io }
     }
 
     #[inline(always)]
@@ -31,22 +57,24 @@ impl {{ctx.type_name|kw_filter}} {
     pub const fn as_ptr(&self) -> *mut () {
         self.ptr.cast::<()>()
     }
+}
 
+impl<IO: peakrdl_rust::io::RegisterIO> {{struct_name}}<'_, IO> {
 {% for reg in ctx.registers %}
     {% set reg_type_name = reg.type_name|kw_filter %}
     {{reg.comment | indent()}}
     #[inline(always)]
     #[must_use]
     {% if reg.array is none %}
-    pub const fn {{reg.inst_name|kw_filter}}(&self) -> peakrdl_rust::reg::Reg<{{reg_type_name}}> {
-        unsafe { peakrdl_rust::reg::Reg::from_ptr(self.ptr.wrapping_byte_add({{"0x{:_X}".format(reg.addr_offset)}}).cast()) }
+    pub const fn {{reg.inst_name|kw_filter}}(&self) -> peakrdl_rust::reg::Reg<'_, {{reg_type_name}}, IO> {
+        unsafe { peakrdl_rust::reg::Reg::from_ptr_with(self.ptr.wrapping_byte_add({{"0x{:_X}".format(reg.addr_offset)}}).cast(), self.io) }
     }
     {% else %}
-    pub const fn {{reg.inst_name|kw_filter}}(&self) -> {{reg.array.type.format("peakrdl_rust::reg::Reg<" ~ reg_type_name ~ ">")}} {
+    pub const fn {{reg.inst_name|kw_filter}}(&self) -> {{reg.array.type.format("peakrdl_rust::reg::Reg<'_, " ~ reg_type_name ~ ", IO>")}} {
         // SAFETY: We will initialize every element before using the array
         let mut array = {{reg.array.type.format("core::mem::MaybeUninit::uninit()")}};
 
-        {% set expr = "unsafe { peakrdl_rust::reg::Reg::<" ~ reg_type_name ~ ">::from_ptr(self.ptr.wrapping_byte_add(" ~ reg.array.addr_offset ~ ").cast()) }"  %}
+        {% set expr = "unsafe { peakrdl_rust::reg::Reg::<'_, " ~ reg_type_name ~ ", IO>::from_ptr_with(self.ptr.wrapping_byte_add(" ~ reg.array.addr_offset ~ ").cast(), self.io) }"  %}
         {{ macros.loop(0, reg.array.dims, expr) | indent(8) }}
 
         // SAFETY: All elements have been initialized above
@@ -58,19 +86,20 @@ impl {{ctx.type_name|kw_filter}} {
 
 {% for node in ctx.submaps %}
     {% set node_type_name = node.type_name|kw_filter %}
+    {% set node_type_name_generics = node_type_name ~ "<'_, IO>" %}
     {{node.comment | indent()}}
     #[inline(always)]
     #[must_use]
     {% if node.array is none %}
-    pub const fn {{node.inst_name|kw_filter}}(&self) -> {{node_type_name}} {
-        unsafe { {{node_type_name}}::from_ptr(self.ptr.wrapping_byte_add({{"0x{:_X}".format(node.addr_offset)}}).cast()) }
+    pub const fn {{node.inst_name|kw_filter}}(&self) -> {{node_type_name_generics}} {
+        unsafe { {{node_type_name}}::from_ptr_with(self.ptr.wrapping_byte_add({{"0x{:_X}".format(node.addr_offset)}}).cast(), self.io) }
     }
     {% else %}
-    pub const fn {{node.inst_name|kw_filter}}(&self) -> {{node.array.type.format(node_type_name)}} {
+    pub const fn {{node.inst_name|kw_filter}}(&self) -> {{node.array.type.format(node_type_name_generics)}} {
         // SAFETY: We will initialize every element before using the array
         let mut array = {{node.array.type.format("core::mem::MaybeUninit::uninit()")}};
 
-        {% set expr = "unsafe { " ~ node_type_name ~ "::from_ptr(self.ptr.wrapping_byte_add(" ~ node.array.addr_offset ~ ").cast()) }"  %}
+        {% set expr = "unsafe { " ~ node_type_name ~ "::<IO>::from_ptr_with(self.ptr.wrapping_byte_add(" ~ node.array.addr_offset ~ ").cast(), self.io) }"  %}
         {{ macros.loop(0, node.array.dims, expr) | indent(8) }}
 
         // SAFETY: All elements have been initialized above
@@ -82,19 +111,20 @@ impl {{ctx.type_name|kw_filter}} {
 
 {% for mem in ctx.memories %}
     {% set mem_type_name = mem.type_name|kw_filter %}
+    {% set mem_type_name_generics = mem_type_name ~ "<'_, IO>" %}
     {{mem.comment | indent()}}
     #[inline(always)]
     #[must_use]
     {% if mem.array is none %}
-    pub const fn {{mem.inst_name|kw_filter}}(&self) -> {{mem_type_name}} {
-        unsafe { {{mem_type_name}}::from_ptr(self.ptr.wrapping_byte_add({{"0x{:_X}".format(mem.addr_offset)}}).cast()) }
+    pub const fn {{mem.inst_name|kw_filter}}(&self) -> {{mem_type_name_generics}} {
+        unsafe { {{mem_type_name}}::from_ptr_with(self.ptr.wrapping_byte_add({{"0x{:_X}".format(mem.addr_offset)}}).cast(), self.io) }
     }
     {% else %}
-    pub const fn {{mem.inst_name|kw_filter}}(&self) -> {{mem.array.type.format(mem_type_name)}} {
+    pub const fn {{mem.inst_name|kw_filter}}(&self) -> {{mem.array.type.format(mem_type_name_generics)}} {
         // SAFETY: We will initialize every element before using the array
         let mut array = {{mem.array.type.format("core::mem::MaybeUninit::uninit()")}};
 
-        {% set expr = "unsafe { " ~ mem_type_name ~ "::from_ptr(self.ptr.wrapping_byte_add(" ~ mem.array.addr_offset ~ ").cast()) }"  %}
+        {% set expr = "unsafe { " ~ mem_type_name ~ "::from_ptr_with(self.ptr.wrapping_byte_add(" ~ mem.array.addr_offset ~ ").cast(), self.io) }"  %}
         {{ macros.loop(0, mem.array.dims, expr) | indent(8) }}
 
         // SAFETY: All elements have been initialized above